backend/requirements.txt
flask
flask-cors
google-genai
gunicorn
openai-whisper
werkzeug
pydub